Refactor Bolt Connection Class
Moved protocol logic into Bolt Class.
